Add vagrant development environment.
This commit adds:

* Vagrant support with virtualbox.
* Ansible playbook for development provisioning.

Most of the things have been generated with phansible. Add some manual customizations though to be compatible with ORO.

After provisioning the VM will have the following:

* nginx
* php5.6 (fpm and cli)
* PostgreSQL 9.3.10

All the modules required for OROcrm to work are provisioned. Sadly you will have to install and configure xdebug (if you want)
on your own. Memory limit is set to 512MB as it is recommended in ORO Documentation.

@see: http://phansible.com/